requirements.txt
accelerate==0.33.0
black==24.8.0
datasets==2.21.0
flake8==7.1.1
hydra-core==1.3.2
isort==5.13.2
jupyter==1.0.0
mypy==1.8.0
neptune==1.11.1
numpy==1.24.3
omegaconf==2.3.0
pre-commit==3.8.0
pydantic==1.10.18
pytest==8.3.2
s3fs==2023.10.0
seaborn==0.13.2
setuptools==65.6.3
torch==2.4.0
transformers==4.36.2
types-PyYAML==6.0.12.20240808
biopython==1.84
fair-esm==2.0.0
h5py==3.11.0
huggingface_hub==0.21.2
lightning==2.1.2
ninja==1.11.1.1
numba==0.60.0
scikit-learn==1.3.2
sequence_models==1.7.0
tape_proteins==0.5
tokenizers==0.15.2
wandb==0.17.7
